Refactored away login.py
